chore: moving the test server around a bit
