+++ package.json
@@ -8,7 +8,7 @@
     "postinstall": "lerna bootstrap --hoist",
     "prettier": "prettier --list-different --write \"./**/**.js\"",
     "publish": "lerna publish",
-    "start": "nodemon testServer.js",
+    "start": "nodemon test/server.js",
     "test": "lerna run test --stream",
     "version": "conventional-changelog --pkg lerna.json -i CHANGELOG.md -s && git add CHANGELOG.md"
   },
