+++ test/payloads/index.js
@@ -1,4 +1,4 @@
 /* eslint-disable global-require */
 module.exports = {
-  standard: require('./har-payloads/standard.har'),
+  standard: require('./standard/har.json'),
 };
